requirements.txt
Pillow==9.4.0
portfolio_carousel==0.1.4
presentation_card==0.1.1
streamlit==1.16.0
